Work on building somethhing sane for shader uniforms.
Changed the way the Canvas::ModelManager works. Accumulating rotations/translations/scale till needed to generate a matrix.
Added some accessor functions to the shader base class allowing the removal of a dynamic pointer cast in the render path.
Fixed a bug with setting the uniforms for textures and multiple shaders.